requirements.txt
requests==2.25.1
uvicorn==0.13.3
starlette==0.13.6
pydantic==1.7.3
fastapi==0.63.0
Pillow==8.1.0
PyJWT==2.0.1
